train functionality

+++ my_project/my_parent/my_service/src/main/java/by/epam/grodno/uladzimir_stsiatsko/my_service/impl/TripListServiceImpl.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Service;
 
 import by.epam.grodno.uladzimir_stsiatsko.my_dao.dao.TripListDao;
 import by.epam.grodno.uladzimir_stsiatsko.my_dao.model.TripList;
+import by.epam.grodno.uladzimir_stsiatsko.my_service.BillService;
 import by.epam.grodno.uladzimir_stsiatsko.my_service.TripListService;
 
 @Service
@@ -13,13 +14,24 @@ public class TripListServiceImpl implements TripListService {
 	@Autowired
 	TripListDao tlDao;
 	
+	@Autowired
+	BillService bService;
+	
 	@Override
 	public void addTripList(TripList tripList){
 		tlDao.insert(tripList);
 	}
 
-	public void deleteTripList(int id){
-		tlDao.deleteTripList(id);
+	public void deleteTripList(int id) throws IllegalArgumentException {
+		if(bService.containsBill(id)){
+			throw new IllegalArgumentException("This trip list can't be deleted because of structural integrity needs");
+		} else {
+			tlDao.deleteTripList(id);
+		}
+	}
+	
+	public boolean containsTrain(int trainId){
+		return tlDao.containsTrain(trainId);
 	}
 	
 }
